[feat] Added sections: Clustering Analysis and Prediction Model
First versions of two new sections.
The Clustering Analysis uses weekly excess mortality data to
compare different countries. The aim is to identify which
government responses were used by countries with the lowest
and higest excess mortality rates.

The Prediction Model uses the LGBM Regressor model to predict
new cases as a time series, then extracts the feature importances
of the model. The aim is to infer which government responses
have had most effect on the growth in the number of cases.